Button Play Homework WIP
Created three buttons that does different functions based on the skeleton of the HTML, CSS, and newly Javascript. Theme: Vaporwave once again. First button changes the color of the header. Second button moves the gif item to the left and back while pulsing. Second button's animation is infinite and alternate. Third is still in debate.